+++ frontend/src/routes/(main)/+layout.svelte
