+++ frontend/src/routes/(main)/searchCourse/+page.svelte
