+++ frontend/src/routes/(main)/studentpage/+page.svelte
